backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
groq>=0.9.0
chromadb>=0.6.0
sentence-transformers>=3.1.0
pymupdf>=1.24.10
pydantic>=2.7.1
python-dotenv>=1.0.1
numpy>=2.0.0
